modified save() method for Item object
The save() method is called when the Item object is being saved to the database.

If the Item object already exists in the database (i.e., it is being updated rather than created), the code checks if the statement field has been changed.

If the statement field has been changed, it creates a new StatementVersion object with the current statement and the statement_updated_at field as the value for the created_at field. This records the time at which the previous text input was replaced by the new one.

After creating the StatementVersion object, it sets the statement_updated_at field to the current time. This updates the time at which the statement field was last modified.

Finally, it calls the parent class's save() method to save the modified Item object to the database.